don't show edit/delete buttons if auth user isn't a note creator

+++ resources/views/book/show.blade.php
@@ -30,13 +30,6 @@
     @if(count($book->articles))
 
         @foreach($book->articles as $article)
-
-            {{--<div class="row">--}}
-                {{--<div class="col-lg-12 col-md-12 col-sm-12 col-xs-12 article-bg">--}}
-                    {{--{!! $article->body !!}--}}
-                {{--</div>--}}
-            {{--</div>--}}
-
             <div class="row">
                 <div class="col-xs-11 col-sm-11 col-md-11 col-lg-11">
                     <div class="article-bg">
@@ -44,46 +37,24 @@
                     </div>
                 </div>
 
+                @if($article->user_id == Auth::id())
+                    <div class="col-xs-1 col-sm-1 col-md-1 col-lg-1" style="padding-left: 0; padding-right: 0">
+                        <div class="btn-group-vertical btn-block btn-group-xs" role="group" aria-label="...">
 
-                <div class="col-xs-1 col-sm-1 col-md-1 col-lg-1" style="padding-left: 0; padding-right: 0">
-                    <div class="btn-group-vertical btn-block btn-group-xs" role="group" aria-label="...">
+                            <a href="{{ action('ArticleController@edit', ['id' => $article->id]) }}" class="btn btn-default">
+                                <i class="fa fa-pencil-square-o fa-2x editIconInLinkGroupButton"></i>
+                            </a>
 
-                        <a href="{{ action('ArticleController@edit', ['id' => $article->id]) }}" class="btn btn-default">
-                            <i class="fa fa-pencil-square-o fa-2x editIconInLinkGroupButton"></i>
-                        </a>
+                            <button type="button" class="btn btn-default hide">
+                                {!! Form::open(['method' => 'DELETE', 'action' => ['ArticleController@destroy', $article->id]]) !!}
 
-                        <button type="button" class="btn btn-default hide">
-                            {!! Form::open(['method' => 'DELETE', 'action' => ['ArticleController@destroy', $article->id]]) !!}
+                                    {!! Form::button('<i class="fa fa-trash-o fa-2x trashIconInColumn"></i>', ['type' => 'submit', 'class' => 'btn btn-default', 'onclick'=>'return confirm("Are you sure?")']) !!}
 
-                                {!! Form::button('<i class="fa fa-trash-o fa-2x trashIconInColumn"></i>', ['type' => 'submit', 'class' => 'btn btn-default', 'onclick'=>'return confirm("Are you sure?")']) !!}
-
-                            {!! Form::close() !!}
-                        </button>
+                                {!! Form::close() !!}
+                            </button>
+                        </div>
                     </div>
-                </div>
-
-
-                {{--<div class="col-xs-1 col-sm-1 col-md-1 col-lg-1 buttonsInColumn">--}}
-                	{{--<div class="row">--}}
-
-                        {{--<div class="col-xs-12 col-sm-12 col-md-12 col-lg-12" style="padding-left: 15px; padding-right: 0">--}}
-                            {{--<a href="{{ action('ArticleController@edit', ['id' => $article->id]) }}">--}}
-                                {{--<i class="fa fa-pencil-square-o fa-2x"></i>--}}
-                            {{--</a>--}}
-                        {{--</div>--}}
-
-                        {{--<div class="col-xs-12 col-sm-12 col-md-12 col-lg-12" style="padding-left: 10px; padding-right: 0">--}}
-                            {{--{!! Form::open(['method' => 'DELETE', 'action' => ['ArticleController@destroy', $article->id]]) !!}--}}
-
-                                {{--{!! Form::button('<i class="fa fa-trash-o fa-2x trashIconInRowInList"></i>', ['type' => 'submit', 'style' => 'border: 0px; background: none;', 'onclick'=>'return confirm("Are you sure?")', 'class' => '']) !!}--}}
-
-                            {{--{!! Form::close() !!}--}}
-                        {{--</div>--}}
-
-                        {{----}}
-
-                    {{--</div>--}}
-                {{--</div>--}}
+                @endif
             </div>
 
             <div class="row">
@@ -92,11 +63,11 @@
                 </div>
             </div>
 
-            <row>
+            <div class="row">
                 <div class="col-xs-11 col-sm-11 col-md-11 col-lg-11">
                     <hr class="hr-list"/>
                 </div>
-            </row>
+            </div>
 
         @endforeach
